requirements-aws.txt
beautifulsoup4==4.12.2
email-validator==2.1.0
flask==3.0.0
flask-caching==2.1.0
flask-login==0.6.3
flask-sqlalchemy==3.1.1
flask-wtf==1.2.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-whois==0.8.0
requests==2.31.0
trafilatura==1.6.3